guided-image-generation/requirements.txt
ipykernel
python-dotenv
openai
requests
streamlit
azure-ai-contentsafety
